Struct ic_crypto_sha256::Sha256

pub struct Sha256 { /* fields omitted */ }

Hasher with fixed algorithm that is guaranteed not to change in the future or across registry versions. The algorithm used to generate the hash is SHA256 and therefore has constant output size of 32 bytes.

This hasher can be used, e.g., for creating fingerprints of files that are persisted on disk.

Example (using state explicitly to hash data piece by piece)

use ic_crypto_sha256::Sha256;

let mut state = Sha256::new();
state.write(b"some ");
state.write(b"data!");
let digest: [u8; 32] = state.finish();

Example (using state implicitly with the convenience function)

use ic_crypto_sha256::Sha256;

let digest: [u8; 32] = Sha256::hash(b"some data!");

Example (using as an std::io::Writer)

use ic_crypto_sha256::Sha256;

let mut reader: &[u8] = b"some data";
let mut hasher = Sha256::new();

std::io::copy(&mut reader, &mut hasher).unwrap();

Implementations

impl Sha256

pub fn new() -> Self

Return a new Sha256 object

pub fn hash(data: &[u8]) -> [u8; 32]

Hashes some data and returns the digest

pub fn write(&mut self, data: &[u8])

Incrementally update the current hash

pub fn finish(self) -> [u8; 32]

Finishes computing a hash, returning the digest